refactor license generation into a new module
